feat: Implement REST APIs for async Kotlin Coroutine example (#1)
Add runnable subproject `provider`:
---
This subproject is responsible for providing data. It consists of a REST API and a H2 in-memory database.

It includes endpoints for fetching:
- `GET /hotels`
- `GET /flights`
- `GET /rentalcars`

These endpoints will return some default dummy data but each endpoint has an artificial delay of 2000 milliseconds to simulate a slow response time.

Add runnable subproject `gateway`:
---
This subproject is responsible for retrieving data from the `provider` REST API and returning it to the client. The interesting part here is that it includes an example of how to correctly utilize Kotlin Coroutines and async/await to execute multiple remote requests. It also includes an example of how it runs without. This allows you to easily notice the difference in performance.

Since each of the 3 requests will take approximately 2000 milliseconds, executing one request at a time will take around 6000 milliseconds to fetch all the data. However, if you use async/await correctly, it can run the requests simultaneously and it should only take around 2000 milliseconds.