test: Add controller tests for gateway subproject (#3)
Add controller tests for gateway subproject:
- Add TravelControllerTestConfig for providing a mocked TravelService using Mockito.
- Add TravelControllerTest for testing that requests to the endpoints return the expected data.
- Add TravelTestData for providing test data.

Since this project is using Spring Boot Webflux, we cannot use the BaseMockMvc class from the spring-boot-kotlin-sample project anymore so it has been deleted. Now we need to use WebTestClient which supports testing Spring Boot Webflux. This specific subproject does not contain a database so we must mock the responses in TravelService.