refac: Rename TravelDetailsDto to TravelDetails
Rename TravelDetailsDto to TravelDetails and move it to the model subproject. The class can be considered a model class and it is pure code that does not use any dependencies so it is fine to place it in the model subproject. We could have separate DTOs if we needed to distinguish between DTOs and models but currently there is no need to do so.