refac: Use WireMock for realistic mocking of HTTP requests (#64)
Update `TravelControllerTest` to use [WireMock](https://github.com/wiremock/wiremock) to realistically mock HTTP requests. Previously, we would mock the method responses of `TravelService` in `TravelControllerTest`. This is one way to execute tests but we are missing a crucial part. The tests had previously no code coverage of `TravelService` but when we use WireMock, the test covers all the code in `TravelService` as well. Additionally, a benefit of using WireMock is that your tests send actual HTTP requests to a live HTTP server with mocked JSON responses.